requirements.txt
django-selectable